+++ test/libsolidity/smtCheckerTests/inheritance/base_contract_assertion_fail_8.sol
@@ -15,4 +15,4 @@ contract C is  A {
 	}
 }
 // ----
-// Warning 6328: (94-108): CHC: Assertion violation happens here.\nCounterexample:\nx = 0\n\nTransaction trace:\nA.constructor()\nState: x = 0\nA.f()
+// Warning 6328: (94-108): CHC: Assertion violation happens here.\nCounterexample:\nx = 0\n\nTransaction trace:\nC.constructor()\nState: x = 0\nA.f()