+++ test/libsolidity/smtCheckerTests/verification_target/no_target_for_constructor_of_abstract_constract.sol
@@ -1,0 +1,7 @@
+pragma experimental SMTChecker;
+
+abstract contract A {
+	constructor() {
+		assert(false); // A cannot be deployed, so this should not be reported
+	}
+}